+++ src/ProcessedStory.ts
@@ -1,79 +1,72 @@
-import * as Parameters from './Parameters';
+import { z as zod } from 'zod';
 import type { StorybookStory } from './browser/StorybookPage';
+
+type Params = {
+  skip: boolean;
+  disabledRules: string[],
+  /** @deprecated */
+  waitForSelector?: string,
+}
 
 /**
  * Story with normalized and custom properties needed by this project.
  */
-export type ProcessedStory = {
-  componentName: string;
+export default class ProcessedStory {
   name: string;
-  parameters: {
-    axe: {
-      skip: boolean;
-      disabledRules: string[];
-      waitForSelector?: string;
-    },
-  };
-  storybookId: string;
+  componentTitle: string;
+  id: string;
+  private parameters: Params;
+
+  constructor(rawStory: StorybookStory) {
+    this.name = rawStory.name;
+    this.componentTitle = rawStory.kind;
+    this.id = rawStory.id;
+    this.parameters = {
+      skip: normalizeSkip(rawStory.parameters?.axe?.skip, rawStory),
+      disabledRules: normalizeDisabledRules(rawStory.parameters?.axe?.disabledRules, rawStory),
+      waitForSelector: normalizeWaitForSelector(rawStory.parameters?.axe?.waitForSelector, rawStory),
+    };
+  }
+
+  get isEnabled() {
+    return !this.parameters.skip;
+  }
+
+  get disabledRules() {
+    return this.parameters.disabledRules;
+  }
+
+  /** @deprecated */
+  get waitForSelector() {
+    return this.parameters.waitForSelector;
+  }
 }
 
-/**
- * Convert a list of Storybook stories into a normalized format.
- */
-export function fromStories(rawStories: StorybookStory[]): ProcessedStory[] {
-  return rawStories.map(fromStory);
-}
-
-/**
- * Convert a raw Storybook story into a normalized format with no optional properties.
- */
-export function fromStory(rawStory: StorybookStory): ProcessedStory {
-  return {
-    componentName: rawStory.kind,
-    name: rawStory.name,
-    parameters: {
-      axe: {
-        skip: normalizeSkip(rawStory.parameters?.axe?.skip, rawStory),
-        disabledRules: normalizeDisabledRules(rawStory.parameters?.axe?.disabledRules, rawStory),
-        waitForSelector: normalizeWaitForSelector(rawStory.parameters?.axe?.waitForSelector, rawStory),
-      },
-    },
-    storybookId: rawStory.id,
-  };
-}
-
-/**
- * Determine if a story is enabled or not.
- */
-export function isEnabled(story: ProcessedStory): boolean {
-  return !story.parameters.axe.skip;
-}
-
-/**
- * Get the Axe rules that a story has disabled.
- */
-export function getDisabledRules(story: ProcessedStory): string[] {
-  return story.parameters.axe.disabledRules;
-}
+const skipSchema = zod.boolean();
+const disabledRulesSchema = zod.array(zod.string());
+const waitForSelectorSchema = zod.optional(zod.string());
 
 function normalizeSkip(skip: unknown, rawStory: StorybookStory) {
   return parseWithFriendlyError(
-    () => Parameters.parseSkip(skip),
-    createInvalidParamErrorMessage(rawStory, 'skip'),
+    () => skipSchema.optional().parse(skip) || false,
+    rawStory,
+    'skip',
   );
 }
 
 function normalizeDisabledRules(disabledRules: unknown, rawStory: StorybookStory) {
   return parseWithFriendlyError(
-    () => Parameters.parseDisabledRules(disabledRules),
-    createInvalidParamErrorMessage(rawStory, 'disabledRules'),
+    () => disabledRulesSchema.optional().parse(disabledRules) || [],
+    rawStory,
+    'disabledRules',
   );
 }
 
 function normalizeWaitForSelector(waitForSelector: unknown, rawStory: StorybookStory) {
   return parseWithFriendlyError(
-    () => Parameters.parseWaitForSelector(waitForSelector),
-    createInvalidParamErrorMessage(rawStory, 'waitForSelector'),
+    () => waitForSelectorSchema.parse(waitForSelector),
+    rawStory,
+    'waitForSelector',
   );
 }
 
@@ -83,22 +76,14 @@ function normalizeWaitForSelector(waitForSelector: unknown, rawStory: StorybookS
  * enough information about what went wrong and where. Instead we'll catch errors from the parsers
  * and re-throw our own.
  */
-function parseWithFriendlyError<T>(parser: () => T, errorMessage: string): T {
+function parseWithFriendlyError<T>(parser: () => T, rawStory: StorybookStory, paramName: string): T {
   try {
     return parser();
   } catch (message) {
-    if (message instanceof Parameters.ParamError) {
-      throw new TypeError(errorMessage);
+    if (message instanceof zod.ZodError) {
+      throw new TypeError(`Invalid value for parameter "${paramName}" in component "${rawStory.kind}", story "${rawStory.name}"`);
     } else {
       throw message;
     }
   }
-}
-
-/**
- * Create useful error text for an invalid param. We provide info on what parameter failed, in
- * which component, and in what story. That way people can easily find their error.
- */
-function createInvalidParamErrorMessage(rawStory: StorybookStory, paramName: string): string {
-  return `Invalid value for parameter "${paramName}" in component "${rawStory.kind}", story "${rawStory.name}"`;
 }