+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 5.0.1 (2022-05-20)
 
+- [breaking] Remove `--format` option
 - [maintenance] Upgrade all dependencies
 - [fix] Access static storybook via file server so everything works with storyStoreV7 (fixing #51)