+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [breaking] Remove `--format` option
 - [fix] Access static storybook via file server so everything works with storyStoreV7 (fixing #51)
+- [fix] Mark `waitForSelector` as deprecated; use play functions instead (https://storybook.js.org/docs/react/writing-stories/play-function)
 - [maintenance] Upgrade all dependencies
 - [maintenance] Use MochaJS internally for running the "tests"
 - [maintenance] Use TypeScript to build the package, instead of esbuild