+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 # Changelog
 
-## 5.0.1 (2022-05-20)
+## Unreleased
 
 - [breaking] Remove `--format` option
-- [fix] Access static storybook via file server so everything works with storyStoreV7 (fixing #51)
+- [fix] Mark `waitForSelector` as deprecated; use play functions instead (https://storybook.js.org/docs/react/writing-stories/play-function)
 - [maintenance] Upgrade all dependencies
 - [maintenance] Use MochaJS internally for running the "tests"
 - [maintenance] Use TypeScript to build the package, instead of esbuild
+
+## 5.0.1 (2022-05-20)
+
+- [fix] Access static storybook via file server so everything works with storyStoreV7 (fixing #51)
+- [maintenance] Upgrade all dependencies
 
 ## 5.0.0 (2021-11-29)